後台電影資訊- 假登入放在共用service
