+++ projects/manager/src/app/pages/movie-page/movie-view-page/movie-view-page.component.ts
@@ -1,10 +1,7 @@
 import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { MoviePageService } from '../services/movie-page.service';
-import { StorageService } from '../../../core/services/storage/storage.service';
 import { MovieDetailRes } from '../../../api/cinePOS-api';
-import { StorageEnum } from '../../../core/enums/storage/storage-enum';
-import { ProfileData } from 'projects/staff/src/app/core/interface/profile-data';
 
 @Component({
   selector: 'app-movie-view-page',
@@ -19,14 +16,13 @@ export class MovieViewPageComponent implements OnInit {
   constructor(
     private _Route: ActivatedRoute,
     private _MoviePageService: MoviePageService,
-    private _StorageService: StorageService,
     private _ChangeDetectorRef: ChangeDetectorRef,
   ) { }
 
   ngOnInit(): void {
     this.movieId = this._Route.snapshot?.params['id'];
     console.log('movieId', this.movieId);
-    this.login();                                              // 登入 (====之後串了真正登入要刪掉)
+    this._MoviePageService.login();                            // 登入 (====之後串了真正登入要刪掉)
     this.getMovieInfoAPI(this.movieId);                        // API- 取得電影資訊
   }
 
@@ -44,17 +40,5 @@ export class MovieViewPageComponent implements OnInit {
     });
   }
 
-
-  // 登入
-  login(): void {
-    let token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0NDRmMGE5OTc3ZmRlZThmYTBiYzc1YSIsInN0YWZmSWQiOiJCMDAwMSIsImlhdCI6MTY4NDY0NzMwMywiZXhwIjoxNjg0OTA2NTAzfQ.6Iv6V2vSUKHS9NJHCInQFDQ9kxTmlN3b9w0zse44Z9U";
-    this._StorageService.setLocalStorage(StorageEnum.token, token);
-    const profileData: ProfileData = {
-      name: "文文檢視頁測試",
-      staffId: "B0001",
-      imgUrl: 'assets/images/angular-icon.webp'
-    }
-    this._StorageService.setLocalStorage(StorageEnum.profileData, profileData);
-  }
 
 }
