+++ projects/manager/src/app/pages/movie-page/services/movie-page.service.ts
@@ -7,6 +7,9 @@ import { CommonUploadSuccess } from '../../../api/cinePOS-api/model/commonUpload
 import { ManagerMovieListSuccess } from '../../../api/cinePOS-api/model/managerMovieListSuccess';
 import { MovieDetailUpdateParameter } from '../../../api/cinePOS-api/model/movieDetailUpdateParameter';
 import { MovieDetailCreateParameterCustomer } from '../../../core/interface/movie';
+import { StorageService } from '../../../core/services/storage/storage.service';
+import { ProfileData } from 'projects/staff/src/app/core/interface/profile-data';
+import { StorageEnum } from '../../../core/enums/storage/storage-enum';
 
 
 @Injectable({
@@ -16,6 +19,7 @@ export class MoviePageService {
 
   constructor(
     private _ManagerService: ManagerService,
+    private _StorageService: StorageService,
   ) { }
 
   // 更新電影資訊
@@ -55,6 +59,19 @@ export class MoviePageService {
         tap(res => res.code !== 1 && alert(res.message)),
         filter(res => res.code === 1)
       )
+  }
+
+
+  // 登入 ====待處理，暫時寫死====
+  login(): void {
+    let token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0NDRmMGE5OTc3ZmRlZThmYTBiYzc1YSIsInN0YWZmSWQiOiJCMDAwMSIsImlhdCI6MTY4NDY5NTMxOSwiZXhwIjoxNjg0OTU0NTE5fQ.4S6yzY0aQgMWFwQc6Y3fnMvzdvU5gxUroqgMD-08yzo";
+    this._StorageService.setLocalStorage(StorageEnum.token, token);
+    const profileData: ProfileData = {
+      name: "文文編輯頁測試",
+      staffId: "B0001",
+      imgUrl: 'assets/images/angular-icon.webp'
+    }
+    this._StorageService.setLocalStorage(StorageEnum.profileData, profileData);
   }
 
 }
